pyproject.toml
[tool.poetry]
name = "cypher-backend"
version = "0.1.0"
description = "Backend for the Cypher application"
authors = ["Your Name <you@example.com>"] # Substitua com suas informações
readme = "README.md" # Crie um README.md se necessário
package-mode = false # Adicionado para indicar que não é um pacote instalável

[tool.poetry.dependencies]
python = "^3.10"
kubernetes = "^32.0.1"
fastapi = {extras = ["standard"], version = "^0.115.12"}
pymongo = "^4.12.0"
pydantic = "^2.11.3"
sse-starlette = "^2.2.1"
docker = "^7.1.0"
pydantic-settings = "^2.9.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cryptography = "^43.0.0"
GitPython = "^3.1.43"
aiofiles = "^23.2.1"
cachetools = "^5.3.3" # Keep for now, or remove if only async-lru is needed
async-lru = "^2.0.4" # Added async-lru

# Adicionado grupo de dependências de desenvolvimento
[tool.poetry.group.dev.dependencies]
sseclient-py = "^1.8.0" # Adicionada dependência para o teste SSE
pytest = "^8.0.0" # Exemplo de outra dependência dev comum
requests = "^2.31.0" # Adicionado requests aqui também, já que o teste usa

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload --host 0.0.0.0 --port 5000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
